requirements.txt
alembic==1.8.1
backend==0.2.4.1
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_Cors==3.0.10
Flask_JWT_Extended==4.6.0
Flask_Mail==0.9.1
Flask_Migrate==3.1.0
flask_sqlalchemy==3.1.1
google_api_python_client==2.143.0
google_auth_oauthlib==1.2.1
protobuf==5.28.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.6
python-dotenv==1.0.1